refactor: useExportConfig shared within PortalPage and its siblings
